backend/requirements.txt
# Financial Leak Detector - Backend Dependencies
# Core Framework
fastapi>=0.109.0
uvicorn>=0.25.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0

# Google APIs (Gmail Integration)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Gemini AI (for smart transaction parsing)
google-generativeai>=0.3.0

# Email & Validation
email-validator>=2.1.0
beautifulsoup4>=4.12.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0
